requirements.txt
accelerate
diffusers
openai-whisper
pyaudio
setuptools
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
transformers